+++ CHANGELOG.md
@@ -101,6 +101,7 @@
 - Fixed the `Always Show Aimable Weapons` feature unintentionally being active when sleeping in a bed.
 - Fixed weapon aiming animations not taking the player's full body rotation into account.
 - Fixed crosshair animations stacking up when user interface is turned off via F11.
+- Fixed issue [#252](https://github.com/Mirsario/TerrariaOverhaul/issues/252) (Spirit Flame & Mystic Coil Snake wrongfully rotated by AlwaysShowAimableWeapons).
 ### Performance
 - Improved performance of falling tree texture generation.